CSE343-Software_Engineering/Term-Project/requirements.txt
django
django-ckeditor
django-crispy-forms
django-cleanup
Pillow
dict2xml
pyjwt
dj-stripe